sb-8 refactor app structure
