+++ src/components/admin-panel.tsx
@@ -9,7 +9,11 @@ import { AdminControls } from "./admin-controls"
 import { AdminAuth } from "./admin-auth"
 import { useCallback } from "react"
 
-export const AdminPanel = () => {
+export type AdminPanelProps = {
+    children: React.ReactNode[];
+}
+
+export const AdminPanel: React.FC<AdminPanelProps> = ({ children }) => {
     const { isAuthenticated } = useAuthentication()
     const { setIsAuthenticated } = useSetAuthentication()
     const { currentSlide } = useSlider()
@@ -20,7 +24,13 @@ export const AdminPanel = () => {
 
     return (
         <>
-            {!isAuthenticated ? <AdminAuth onAuthenticated={onAuthenticated} /> : <AdminControls />}
+            {!isAuthenticated ? (
+                <AdminAuth onAuthenticated={onAuthenticated} />
+            ) : (
+                <AdminControls>
+                    {children}
+                </AdminControls>
+            )}
 
             <Link
                 href={`/${currentSlide}`}
